Fixes #3 added confirm dialog when deleting player

+++ src/app/players/players.component.ts
@@ -1,7 +1,10 @@
 import { Component, OnInit } from '@angular/core';
 
+import { MdDialogRef, MdDialog } from '@angular/material';
+
 import { GameService } from '../game.service';
 import { Player } from '../player';
+import { ConfirmDialogComponent } from '../confirm-dialog/confirm-dialog.component';
 
 @Component({
   selector: 'app-players',
@@ -12,8 +15,9 @@ export class PlayersComponent implements OnInit {
 
   player: Player = null;
   players: Player[];
+  dialogRef: MdDialogRef<ConfirmDialogComponent>;
 
-  constructor(private gameService: GameService) { }
+  constructor(private gameService: GameService, public dialog: MdDialog) { }
 
   getPlayers(): void {
     this.gameService.getPlayers().then(players => { 
@@ -26,7 +30,16 @@ export class PlayersComponent implements OnInit {
   }
 
   onDelete(player: Player): void {
-    this.gameService.removePlayer(player);
+    this.dialogRef = this.dialog.open(ConfirmDialogComponent, {
+      disableClose: false
+    });
+
+    this.dialogRef.afterClosed().subscribe(result => {
+      if ('yes' === result) {
+        this.gameService.removePlayer(player);
+      }
+      this.dialogRef = null;
+    });
   }
 
   onAdd(): void {
